ms-server/requirements.txt
fastapi[all]
uvicorn
python-jose
peewee
python-dotenv
pydantic
pytz
torch
openai-whisper
dataclasses-json
aenum